requirements.txt
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.2.0
plotly==5.18.0
